requirements.txt
fastapi[standard]
torch
python-dotenv
numpy
requests
aiofiles
python-json-logger